frontend/app/requirements.txt
dash
dash-bootstrap-components
plotly
pandas
numpy
psycopg2-binary
sqlalchemy